exchange screen

+++ app/src/main/java/com/vm/revoluttest/domain/gson/CurrencyEnumMapDeserializer.java
@@ -8,18 +8,22 @@ import com.google.gson.reflect.TypeToken;
 import com.vm.revoluttest.domain.model.Currency;
 
 import java.lang.reflect.Type;
+import java.math.BigDecimal;
 import java.util.EnumMap;
 import java.util.Map;
 
-public class CurrencyEnumMapDeserializer implements JsonDeserializer<EnumMap<Currency, Double>> {
+/**
+ * deserializer that maps Map<String, BigDecimal> into EnumMap<Currency, BigDecimal>
+ */
+public class CurrencyEnumMapDeserializer implements JsonDeserializer<EnumMap<Currency, BigDecimal>> {
     @Override
-    public EnumMap<Currency, Double> deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+    public EnumMap<Currency, BigDecimal> deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
             throws JsonParseException {
-        Type mapType = new TypeToken<Map<String, Double>>() {}.getType();
-        Map<String, Double> rates= context.deserialize(json,mapType);
-        EnumMap<Currency, Double> convertedRates = new EnumMap<>(Currency.class);
+        Type mapType = new TypeToken<Map<String, BigDecimal>>() {}.getType();
+        Map<String, BigDecimal> rates= context.deserialize(json,mapType);
+        EnumMap<Currency, BigDecimal> convertedRates = new EnumMap<>(Currency.class);
         for (Currency currency : Currency.values()) {
-            Double rate = rates.get(currency.getName());
+            BigDecimal rate = rates.get(currency.getName());
             if(rate!=null){
                 convertedRates.put(currency, rate);
             }
